+++ includes/structure/layout.php
@@ -9,5 +9,5 @@ if( !defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
  *
  * @since 2.3.38
  */
-// add_filter( 'genesis_pre_get_option_site_layout', '__genesis_return_full_width_content' );
+// add_filter( 'genesis_site_layout', '__genesis_return_full_width_content' );
 // Other possible layouts: __genesis_return_content_sidebar, __genesis_return_sidebar_content, __genesis_return_content_sidebar_sidebar, __genesis_return_sidebar_sidebar_content, __genesis_return_sidebar_content_sidebar, __genesis_return_full_width_content